Frotnend changes
Fully integrated Firebase authentication with a login page/sign up page both with basic logins and with google provider, created a UI for the login/signups. Adjust the Landing pages CSS/Layout to be modulerized and propertly layed out so that footer is not hard coded in the home page.